Updating pk_individual plots

+++ config/pk_individual.py
@@ -57,6 +57,7 @@ if __name__ == "__main__":
         fitter.fit(file)
 
     if fitter.should_plot():
+        from os import path
         import matplotlib.pyplot as plt
 
         import logging
@@ -64,42 +65,54 @@ if __name__ == "__main__":
         logging.info("Creating plots")
 
         res = {}
-        for posterior, weight, chain, evidence, model, data, extra in fitter.load():
-            n = extra["name"].split(",")[0]
-            if res.get(n) is None:
-                res[n] = []
-            i = posterior.argmax()
-            chi2 = -2 * posterior[i]
-            m, s = weighted_avg_and_std(chain[:, 0], weight)
-            print(m)
-            res[n].append([m, s, chain[i, 0], posterior[i], chi2, -chi2, extra["realisation"], evidence.max()])
-        for label in res.keys():
-            res[label] = pd.DataFrame(res[label], columns=["avg", "std", "max", "posterior", "chi2", "Dchi2", "realisation", "evidence"])
+        if path.exists(pfn + "_alphameans.csv"):
+            logging.info("Found alphameans.csv, reading from existing file")
 
-        ks = list(res.keys())
-        all_ids = pd.concat(tuple([res[l][["realisation"]] for l in ks]))
-        counts = all_ids.groupby("realisation").size().reset_index()
-        max_count = counts.values[:, 1].max()
-        good_ids = counts.loc[counts.values[:, 1] == max_count, ["realisation"]]
+            df = pd.read_csv(pfn + "_alphameans.csv")
+            labels = [c.replace("_pk_mean", "") for c in df.columns if "_pk_mean" in c]
+            for label in labels:
+                res[label] = df[[label + "_pk_mean", label + "_pk_std", "realisation", label + "_pk_evidence"]].copy()
+                res[label].rename({label + "_pk_mean": "avg", label + "_pk_std": "std", label + "_pk_evidence": "evidence"}, axis="columns", inplace=True)
+        else:
 
-        for label, df in res.items():
-            res[label] = pd.merge(good_ids, df, how="left", on="realisation")
+            logging.info("Didn't find alphameans.csv, reading chains")
 
-        df_all = None
-        for label, means in res.items():
-            d = pd.DataFrame(
-                {
-                    "realisation": means["realisation"],
-                    f"{label}_pk_mean": means["avg"],
-                    f"{label}_pk_std": means["std"],
-                    f"{label}_pk_evidence": means["evidence"],
-                }
-            )
-            if df_all is None:
-                df_all = d
-            else:
-                df_all = pd.merge(df_all, d, how="outer", on="realisation")
-        df_all.to_csv(pfn + "_alphameans.csv", index=False, float_format="%0.5f")
+            for posterior, weight, chain, evidence, model, data, extra in fitter.load():
+                n = extra["name"].split(",")[0]
+                if res.get(n) is None:
+                    res[n] = []
+                i = posterior.argmax()
+                chi2 = -2 * posterior[i]
+                m, s = weighted_avg_and_std(chain[:, 0], weight)
+                print(m)
+                res[n].append([m, s, chain[i, 0], posterior[i], chi2, -chi2, extra["realisation"], evidence.max()])
+            for label in res.keys():
+                res[label] = pd.DataFrame(res[label], columns=["avg", "std", "max", "posterior", "chi2", "Dchi2", "realisation", "evidence"])
+
+            ks = list(res.keys())
+            all_ids = pd.concat(tuple([res[l][["realisation"]] for l in ks]))
+            counts = all_ids.groupby("realisation").size().reset_index()
+            max_count = counts.values[:, 1].max()
+            good_ids = counts.loc[counts.values[:, 1] == max_count, ["realisation"]]
+
+            for label, df in res.items():
+                res[label] = pd.merge(good_ids, df, how="left", on="realisation")
+
+            df_all = None
+            for label, means in res.items():
+                d = pd.DataFrame(
+                    {
+                        "realisation": means["realisation"],
+                        f"{label}_pk_mean": means["avg"],
+                        f"{label}_pk_std": means["std"],
+                        f"{label}_pk_evidence": means["evidence"],
+                    }
+                )
+                if df_all is None:
+                    df_all = d
+                else:
+                    df_all = pd.merge(df_all, d, how="outer", on="realisation")
+            df_all.to_csv(pfn + "_alphameans.csv", index=False, float_format="%0.5f")
 
         # Define colour scheme
         c2 = ["#225465", "#5FA45E"]  # ["#581d7f", "#e05286"]
@@ -108,15 +121,14 @@ if __name__ == "__main__":
         c5 = ["#262232", "#1F4D5C", "#0E7A6E", "#5BA561", "#C1C64D"]  # ["#3c1357", "#61208d", "#a73b8f", "#e8638b", "#f4aea3"]
         cols = {"Beutler": c4[0], "Seo": c4[1], "Ding": c4[2], "Noda": c4[3]}
 
-        # plt.rc("text", usetex=True)
-        # plt.rc("font", family="serif")
+        # Bins for means
         bins_both = np.linspace(0.91, 1.08, 31)
         bins = np.linspace(0.95, 1.06, 31)
         ticks = [0.97, 1.0, 1.03]
         lim = bins[0], bins[-1]
         lim_both = bins_both[0], bins_both[-1]
 
-        # Make histogram comparison
+        # Make histogram comparison of the means
         if True:
             fig, axes = plt.subplots(nrows=2, figsize=(5, 4), sharex=True)
             for label, means in res.items():
@@ -134,9 +146,9 @@ if __name__ == "__main__":
             axes[1].axvline(1.0, color="k", lw=1, ls="--", alpha=0.6)
             axes[0].annotate("Prerecon", (0.98, 0.96), xycoords="axes fraction", horizontalalignment="right", verticalalignment="top")
             axes[1].annotate("Recon", (0.98, 0.96), xycoords="axes fraction", horizontalalignment="right", verticalalignment="top")
-            # leg1 = axes[0].legend(loc=2, frameon=False)
-            leg2 = axes[1].legend(loc=2, frameon=False)
-            for lh in leg2.legendHandles:
+            leg1 = axes[0].legend(loc=2, frameon=False)
+            # leg2 = axes[1].legend(loc=2, frameon=False)
+            for lh in leg1.legendHandles:
                 lh.set_alpha(1)
             axes[0].tick_params(axis="y", left=False)
             axes[1].tick_params(axis="y", left=False)
@@ -146,45 +158,19 @@ if __name__ == "__main__":
             fig.savefig(pfn + "_alphahist.png", bbox_inches="tight", dpi=300, transparent=True)
             fig.savefig(pfn + "_alphahist.pdf", bbox_inches="tight", dpi=300, transparent=True)
 
-        # Make histogram comparison
-        if True:
-            fig, axes = plt.subplots(nrows=2, figsize=(5, 4), sharex=True)
-            for label, means in res.items():
-                if "Smooth" in label:
-                    continue
-                if "Prerecon" in label:
-                    ax = axes[0]
-                else:
-                    ax = axes[1]
-                c = cols[label.split()[0]]
-                ax.hist(means["std"], bins=bins_both, label=label, histtype="stepfilled", linewidth=2, alpha=0.3, color=c)
-                ax.hist(means["std"], bins=bins_both, histtype="step", linewidth=1.5, color=cols[label.split()[0]])
-            axes[1].set_xlabel(r"$\langle \sigma_{\alpha} \rangle$", fontsize=14)
-            axes[0].set_yticklabels([])
-            axes[1].set_yticklabels([])
-            # axes[0].axvline(1.0, color="k", lw=1, ls="--", alpha=0.6)
-            # axes[1].axvline(1.0, color="k", lw=1, ls="--", alpha=0.6)
-            leg1 = axes[0].legend(loc=1, frameon=False)
-            leg2 = axes[1].legend(loc=1, frameon=False)
-            for lh in leg1.legendHandles + leg2.legendHandles:
-                lh.set_alpha(1)
-            # axes[0].tick_params(axis='y', left=False)
-            # axes[1].tick_params(axis='y', left=False)
-            plt.subplots_adjust(hspace=0.0)
-            fig.savefig(pfn + "_alphaerrhist.png", bbox_inches="tight", dpi=300, transparent=True)
-            fig.savefig(pfn + "_alphaerrhist.pdf", bbox_inches="tight", dpi=300, transparent=True)
-
-        from matplotlib.colors import to_rgb, to_hex
-
-        def blend_hex(hex1, hex2):
-            a = np.array(to_rgb(hex1))
-            b = np.array(to_rgb(hex2))
-            return to_hex(0.5 * (a + b))
-
         # Alpha-alpha comparison
         if True:
+
+            from matplotlib.colors import to_rgb, to_hex
+
+            def blend_hex(hex1, hex2):
+                a = np.array(to_rgb(hex1))
+                b = np.array(to_rgb(hex2))
+                return to_hex(0.5 * (a + b))
+
             from scipy.interpolate import interp1d
 
+            # Post-recon
             cols = {"Beutler": c4[0], "Seo": c4[1], "Ding": c4[2], "Noda": c4[3]}
             fig, axes = plt.subplots(5, 5, figsize=(10, 10), sharex=True)
             labels = ["Beutler 2017 Recon", "Beutler 2017 Fixed $\\Sigma_{nl}$ Recon", "Seo 2016 Recon", "Ding 2018 Recon", "Noda 2019 Recon"]
@@ -237,6 +223,7 @@ if __name__ == "__main__":
             fig.savefig(pfn + "_alphacomp.png", bbox_inches="tight", dpi=300, transparent=True)
             fig.savefig(pfn + "_alphacomp.pdf", bbox_inches="tight", dpi=300, transparent=True)
 
+            # Pre-recon
             fig, axes = plt.subplots(5, 5, figsize=(10, 10), sharex=True)
             labels = ["Beutler 2017 Prerecon", "Beutler 2017 Fixed $\\Sigma_{nl}$ Prerecon", "Seo 2016 Prerecon", "Ding 2018 Prerecon", "Noda 2019 Prerecon"]
             for i, label1 in enumerate(labels):
@@ -287,14 +274,50 @@ if __name__ == "__main__":
             fig.savefig(pfn + "_alphacomp_prerecon.png", bbox_inches="tight", dpi=300, transparent=True)
             fig.savefig(pfn + "_alphacomp_prerecon.pdf", bbox_inches="tight", dpi=300, transparent=True)
 
+        # Bins for the alpha errors
+        bins_both = np.linspace(0.005, 0.04, 31)
+        lim_both = bins_both[0], bins_both[-1]
+
+        # Make histogram comparison of the errors
+        if True:
+            fig, axes = plt.subplots(nrows=2, figsize=(5, 4), sharex=True)
+            for label, means in res.items():
+                if "Prerecon" in label:
+                    ax = axes[0]
+                else:
+                    ax = axes[1]
+                c = cols[label.split()[0]]
+                ax.hist(means["std"], bins=bins_both, label=" ".join(label.split()[:-1]), histtype="stepfilled", linewidth=2, alpha=0.3, color=c)
+                ax.hist(means["std"], bins=bins_both, histtype="step", linewidth=1.5, color=cols[label.split()[0]])
+            axes[1].set_xlabel(r"$\sigma_{\alpha}$", fontsize=14)
+            axes[0].set_yticklabels([])
+            axes[1].set_yticklabels([])
+            axes[0].annotate("Prerecon", (0.98, 0.96), xycoords="axes fraction", horizontalalignment="right", verticalalignment="top")
+            axes[1].annotate("Recon", (0.98, 0.96), xycoords="axes fraction", horizontalalignment="right", verticalalignment="top")
+            # leg1 = axes[0].legend(loc=1, frameon=False)
+            leg2 = axes[1].legend(loc=4, frameon=False)
+            for lh in leg2.legendHandles:
+                lh.set_alpha(1)
+            axes[0].tick_params(axis="y", left=False)
+            axes[1].tick_params(axis="y", left=False)
+            axes[0].set_xlim(*lim_both)
+            axes[1].set_xlim(*lim_both)
+            plt.subplots_adjust(hspace=0.0)
+            fig.savefig(pfn + "_alphaerrhist.png", bbox_inches="tight", dpi=300, transparent=True)
+            fig.savefig(pfn + "_alphaerrhist.pdf", bbox_inches="tight", dpi=300, transparent=True)
+
+        # Error-error comparison
         if True:
             from scipy.interpolate import interp1d
 
-            bins = np.linspace(0.02, 0.17, 31)
+            # Post-recon
+            bins = np.linspace(0.004, 0.02, 31)
+            ticks = [0.008, 0.012, 0.016]
+            lim = bins[0], bins[-1]
+
             cols = {"Beutler": c4[0], "Seo": c4[1], "Ding": c4[2], "Noda": c4[3]}
-            fig, axes = plt.subplots(4, 4, figsize=(10, 10), sharex=True)
+            fig, axes = plt.subplots(5, 5, figsize=(10, 10), sharex=True)
             labels = ["Beutler 2017 Recon", "Beutler 2017 Fixed $\\Sigma_{nl}$ Recon", "Seo 2016 Recon", "Ding 2018 Recon", "Noda 2019 Recon"]
-            v1, v2 = 0.01, 0.17
             k = "std"
             for i, label1 in enumerate(labels):
                 for j, label2 in enumerate(labels):
@@ -307,44 +330,46 @@ if __name__ == "__main__":
                         ax.hist(res[label1][k], bins=bins, histtype="step", linewidth=1.5, color=cols[label1.split()[0]])
                         ax.set_yticklabels([])
                         ax.tick_params(axis="y", left=False)
-                        ax.set_xlim(v1, v2)
+                        ax.set_xlim(*lim)
                         ax.spines["right"].set_visible(False)
                         ax.spines["top"].set_visible(False)
                         if j == 0:
                             ax.spines["left"].set_visible(False)
-                        if j == 3:
-                            ax.set_xlabel(label2, fontsize=12)
-                            # ax.set_xticks([0.9, 1.0, 1.1])
+                        if j == 4:
+                            ax.set_xlabel(" ".join(label2.split()[:-1]), fontsize=12)
+                            ax.set_xticks(ticks)
                     else:
                         print(label1, label2)
-                        a1 = np.array(res[label2][:, 1])
-                        a2 = np.array(res[label1][:, 1])
-                        c = blend_hex(cols[label1.split()[0]], cols[label2.split()[0]])
+                        a1 = np.array(res[label2][k])
+                        a2 = np.array(res[label1][k])
                         c = np.abs(a1 - a2)
-                        ax.scatter(a1, a2, s=2, c=c, cmap="viridis_r", vmin=-0.01, vmax=0.15)
-                        ax.set_xlim(v1, v2)
-                        ax.set_ylim(v1, v2)
-                        ax.plot([v1, v2], [v1, v2], c="k", lw=1, alpha=0.8, ls=":")
-                        # ax.axvline(1.0, color="k", lw=1, ls="--", alpha=0.4)
-                        # ax.axhline(1.0, color="k", lw=1, ls="--", alpha=0.4)
+                        # ax.scatter(a1, a2, s=2, c=c, cmap="viridis_r", vmin=-0.01, vmax=0.15)
+                        ax.scatter(a1, a2, s=2, c=c, cmap="viridis_r", vmin=-0.0001, vmax=0.005)
+                        ax.set_xlim(*lim)
+                        ax.set_ylim(*lim)
+                        ax.plot([0.0, 1.0], [0.0, 1.0], c="k", lw=1, alpha=0.8, ls=":")
 
                         if j != 0:
                             ax.set_yticklabels([])
                             ax.tick_params(axis="y", left=False)
                         else:
-                            ax.set_ylabel(label1, fontsize=12)
-                            # ax.set_yticks([0.9, 1.0, 1.1])
-                        if i == 3:
-                            ax.set_xlabel(label2, fontsize=12)
-                            # ax.set_xticks([0.9, 1.0, 1.1])
+                            ax.set_ylabel(" ".join(label1.split()[:-1]), fontsize=12)
+                            ax.set_yticks(ticks)
+                        if i == 4:
+                            ax.set_xlabel(" ".join(label2.split()[:-1]), fontsize=12)
+                            ax.set_xticks(ticks)
             plt.subplots_adjust(hspace=0.0, wspace=0)
             fig.savefig(pfn + "_alphaerrcomp.png", bbox_inches="tight", dpi=300, transparent=True)
             fig.savefig(pfn + "_alphaerrcomp.pdf", bbox_inches="tight", dpi=300, transparent=True)
 
-            bins = np.linspace(0.02, 0.17, 31)
-            fig, axes = plt.subplots(4, 4, figsize=(10, 10), sharex=True)
+            # Pre-recon
+            bins = np.linspace(0.005, 0.045, 31)
+            ticks = [0.01, 0.025, 0.04]
+            lim = bins[0], bins[-1]
+
+            fig, axes = plt.subplots(5, 5, figsize=(10, 10), sharex=True)
             labels = ["Beutler 2017 Prerecon", "Beutler 2017 Fixed $\\Sigma_{nl}$ Prerecon", "Seo 2016 Prerecon", "Ding 2018 Prerecon", "Noda 2019 Prerecon"]
-            v1, v2 = 0.01, 0.17
+            k = "std"
             for i, label1 in enumerate(labels):
                 for j, label2 in enumerate(labels):
                     ax = axes[i, j]
@@ -356,36 +381,34 @@ if __name__ == "__main__":
                         ax.hist(res[label1][k], bins=bins, histtype="step", linewidth=1.5, color=cols[label1.split()[0]])
                         ax.set_yticklabels([])
                         ax.tick_params(axis="y", left=False)
-                        ax.set_xlim(v1, v2)
+                        ax.set_xlim(*lim)
                         ax.spines["right"].set_visible(False)
                         ax.spines["top"].set_visible(False)
                         if j == 0:
                             ax.spines["left"].set_visible(False)
-                        if j == 3:
-                            ax.set_xlabel(label2, fontsize=12)
-                            # ax.set_xticks([0.9, 1.0, 1.1])
+                        if j == 4:
+                            ax.set_xlabel(" ".join(label2.split()[:-1]), fontsize=12)
+                            ax.set_xticks(ticks)
                     else:
                         print(label1, label2)
-                        a1 = np.array(res[label2][:, 1])
-                        a2 = np.array(res[label1][:, 1])
-                        c = blend_hex(cols[label1.split()[0]], cols[label2.split()[0]])
+                        a1 = np.array(res[label2][k])
+                        a2 = np.array(res[label1][k])
                         c = np.abs(a1 - a2)
-                        ax.scatter(a1, a2, s=2, c=c, cmap="viridis_r", vmin=-0.01, vmax=0.15)
-                        ax.set_xlim(v1, v2)
-                        ax.set_ylim(v1, v2)
-                        ax.plot([v1, v2], [v1, v2], c="k", lw=1, alpha=0.8, ls=":")
-                        # ax.axvline(1.0, color="k", lw=1, ls="--", alpha=0.4)
-                        # ax.axhline(1.0, color="k", lw=1, ls="--", alpha=0.4)
+                        # ax.scatter(a1, a2, s=2, c=c, cmap="viridis_r", vmin=-0.01, vmax=0.15)
+                        ax.scatter(a1, a2, s=2, c=c, cmap="viridis_r", vmin=-0.0001, vmax=0.01)
+                        ax.set_xlim(*lim)
+                        ax.set_ylim(*lim)
+                        ax.plot([0.0, 1.0], [0.0, 1.0], c="k", lw=1, alpha=0.8, ls=":")
 
                         if j != 0:
                             ax.set_yticklabels([])
                             ax.tick_params(axis="y", left=False)
                         else:
-                            ax.set_ylabel(label1, fontsize=12)
-                            # ax.set_yticks([0.9, 1.0, 1.1])
-                        if i == 3:
-                            ax.set_xlabel(label2, fontsize=12)
-                            # ax.set_xticks([0.9, 1.0, 1.1])
+                            ax.set_ylabel(" ".join(label1.split()[:-1]), fontsize=12)
+                            ax.set_yticks(ticks)
+                        if i == 4:
+                            ax.set_xlabel(" ".join(label2.split()[:-1]), fontsize=12)
+                            ax.set_xticks(ticks)
             plt.subplots_adjust(hspace=0.0, wspace=0)
             fig.savefig(pfn + "_alphaerrcomp_prerecon.png", bbox_inches="tight", dpi=300, transparent=True)
             fig.savefig(pfn + "_alphaerrcomp_prerecon.pdf", bbox_inches="tight", dpi=300, transparent=True)
